multi-vendor-more-core
